[Proof-of-concept] Buggy hotkeys directive
Summary:
More a proof-of-concept and future code reference than anything else.

You can add hotkeys with a "hotkey=" directive.  The element must be
an <a href=> or have an ng-click directive.  If you hold down command,
the components which have an associated hotkey will be hilighted with
that component.

Test Plan: None

Reviewers: bengotow

Reviewed By: bengotow

Differential Revision: https://review.inboxapp.com/D411